backend/requirements.txt
asgiref==3.3.4
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.5.30
cffi==1.17.1
charset-normalizer==3.3.2
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==43.0.1
defusedxml==0.8.0rc2
Django==3.2.4
django-autoslug==1.9.8
django-debug-toolbar==3.2.1
django-environ==0.4.5
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-spectacular==0.27.2
drf-yasg==1.20.0
exceptiongroup==1.2.2
flake8==3.9.2
flake8-broken-line==0.6.0
flake8-plugin-utils==1.3.2
flake8-polyfill==1.0.2
flake8-return==1.1.3
gunicorn==20.1.0
idna==3.10
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.4
jsonschema==4.17.3
MarkupSafe==2.1.5
mccabe==0.6.1
oauthlib==3.2.2
packaging==20.9
pdfkit==0.6.1
pillow==10.4.0
pluggy==0.13.1
psycopg2-binary==2.9.1
py==1.10.0
pycodestyle==2.7.0
pycparser==2.22
pyflakes==2.3.1
PyJWT==2.9.0
pyparsing==2.4.7
pyrsistent==0.20.0
pytest==7.4.4
python-dotenv==0.17.1
python-slugify==5.0.2
python3-openid==3.2.0
pytz==2021.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
short_url==1.2.2
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sorl-thumbnail==12.7.0
sqlparse==0.4.1
testfixtures==6.17.1
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==1.26.6
users==1.0.dev0
